+++ src/help.rs
@@ -18,12 +18,12 @@ USAGE:
 FLAGS:
     --help                       # Print this message and exit
     --version                    # Print version string and exit
+    --watch                      # Rerun tests on file changes
     --compiler /path/to/compiler # Precis the compiler to use (defaults to just elm)
     --seed integer               # Run with initial fuzzer seed (defaults to random)
     --fuzz integer               # Precise number of iterations of fuzz tests (defaults to 100)
     --workers integer            # Precise number of worker threads (defaults to number of logic cores)
     --report console|json|junit  # Print results to stdout in given format (defaults to console)
-    (--watch)                    # (Not supported yet) Run tests on file changes
 
 SUBCOMMANDS:
     init               # Initialize tests dependencies and directory